update renamed imports
