+++ README.md
@@ -117,7 +117,7 @@ As it can be a bit difficult to follow, generally this is doing something like.
 For example, the JSON normaliser calls a `cleanup` function before it starts processing the data. If you wanted to remove the `images` key like shown above, you could do so like:
 
 ```python
-from bleanser.modules.json_new import JsonNormaliser, delkeys, Json
+from bleanser.core.modules.json_new import JsonNormaliser, delkeys, Json
 
 
 class Normaliser(JsonNormaliser):
@@ -136,8 +136,8 @@ if __name__ == '__main__':
 For common formats, the helper classes handle all the tedious bits like loading/parsing data, managing the temporary files. The `Normaliser.main` calls the CLI, which looks like this:
 
 ```
- $ python3 -m bleanser.modules.json_new prune --help
-Usage: python -m bleanser.modules.json_new prune [OPTIONS] PATH
+ $ python3 -m bleanser.core.modules.json_new prune --help
+Usage: python -m bleanser.core.modules.json_new prune [OPTIONS] PATH
 
 Options:
   --glob                 Treat the path as glob (in the glob.glob sense)
@@ -164,7 +164,7 @@ import json
 from pathlib import Path
 from typing import Iterator
 
-from bleanser.modules.extract import ExtractObjectsNormaliser
+from bleanser.core.modules.extract import ExtractObjectsNormaliser
 
 
 class Normaliser(ExtractObjectsNormaliser):
