+++ src/bleanser/modules/monzo.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-from bleanser.modules.json_new import JsonNormaliser, Json, delkeys
+from bleanser.core.modules.json_new import JsonNormaliser, Json, delkeys
 
 
 class Normaliser(JsonNormaliser):
