+++ src/bleanser/modules/rescuetime.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-from bleanser.modules.json_new import JsonNormaliser
+from bleanser.core.modules.json_new import JsonNormaliser
 
 
 class Normaliser(JsonNormaliser):
